requirements.txt
nltk
scikit-learn
numpy
customtkinter
Pillow